tarefa-convensao-de-codigo

+++ src/main/java/br/com/melo/api/services/ProductService.java
@@ -57,7 +57,7 @@ public class ProductService {
                 nota += review.getRate();
             }
 
-            nota /= (double)product.getReviews().size();
+            nota /= product.getReviews().size();
         }
         product.setReviewRate(nota);
         return product;
